requirements.txt
flask
flask-cors
Pyrebase4
firebase-admin
google-generativeai
requests
gunicorn
cachecontrol
